requirements.txt
numpy
pandas
scikit-learn
torch
tqdm
matplotlib
umap-learn
sentence-transformers
joblib
